+++ next-docs/pages/components/listItem.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import Preview from '../../components/codePreview/Preview';
+import PropsTable from "../../components/PropsTable";
 import SingleLineItem from '../../public/examples/listItem/SingleLineItem';
 import SingleLineItemIcon from '../../public/examples/listItem/SingleLineItemIcon';
 import SingleLineItemSize from '../../public/examples/listItem/SingleLineItemSize';
@@ -75,6 +76,61 @@ export default function PageListItem() {
           code={examples ? examples.TwoLineItemsCheckbox : 'Loading'}
         />
       </section>
+
+      <PropsTable
+        title="Props"
+        data={[
+          {
+            name: 'subtext',
+            type: 'React.ReactElement',
+            required: false,
+            default: '-',
+            description: 'List item subtext',
+          },
+          {
+            name: 'size',
+            type: 'medium | large',
+            required: false,
+            default: '-',
+            description: 'List item size',
+          },
+          {
+            name: 'backgroundColor',
+            type: 'ColorProps',
+            required: false,
+            default: '-',
+            description: 'List item background color',
+          },
+          {
+            name: 'color',
+            type: 'ColorProps',
+            required: false,
+            default: '-',
+            description: 'List item color',
+          },
+          {
+            name: 'elementLeft',
+            type: 'React.ReactElement',
+            required: false,
+            default: '-',
+            description: 'Element on the left',
+          },
+          {
+            name: 'elementRight',
+            type: 'React.ReactElement',
+            required: false,
+            default: '-',
+            description: 'Element on the right',
+          },
+          {
+            name: 'isMeta',
+            type: 'boolean',
+            required: false,
+            default: 'false',
+            description: '-',
+          },
+        ]}
+      />
     </>
   );
 }