+++ next-docs/public/examples/checkbox/Readonly.tsx
@@ -1,7 +1,5 @@
 import { Checkbox } from '@heathmont/moon-core';
 
-const Example = () => (
-  <Checkbox readOnly={true} label="Readonly checkbox." />
-);
+const Example = () => <Checkbox readOnly label="Readonly checkbox" />;
 
 export default Example;