+++ packages/core/src/avatar/private/types/SizeProps.ts
@@ -1,0 +1,17 @@
+import Size from '../../../private/enums/Size';
+
+type SizeProps =
+  | Size.XSMALL
+  | Size.SMALL
+  | Size.MEDIUM
+  | Size.LARGE
+  | Size.XLARGE
+  | Size.TWOXLARGE
+  | 'xsmall'
+  | 'small'
+  | 'medium'
+  | 'large'
+  | 'xlarge'
+  | 'twoxlarge';
+
+export default SizeProps;