+++ packages/core/src/checkbox/Checkbox.tsx
@@ -1,4 +1,4 @@
-import React, {forwardRef, useState} from 'react';
+import React, { forwardRef, useState } from 'react';
 import CheckboxIconWrapper from './styles/CheckboxIconWrapper';
 import CheckboxInput from './styles/CheckboxInput';
 import CheckboxLabel from './styles/CheckboxLabel';
@@ -17,33 +17,43 @@ export interface CheckboxProps {
 }
 
 const Checkbox = forwardRef<HTMLInputElement, CheckboxProps>(
-  ({ disabled, readOnly, ariaLabel, label, id, variant, dir, ...inputProps }, ref) => {
-    const [isChecked, setIsChecked] = useState(false)
+  (
+    { disabled, readOnly, ariaLabel, label, id, variant, dir, ...inputProps },
+    ref
+  ) => {
+    const [isChecked, setIsChecked] = useState(false);
 
-    return (<CheckboxLabel htmlFor={id} variant={variant}>
-      <CheckboxInput
-        id={id}
-        disabled={disabled}
-        aria-label={ariaLabel}
-        ref={ref}
-        {...inputProps}
-        readOnly={readOnly}
-        onClick={(e) => {
-          if (disabled || readOnly) {
-            e.preventDefault();
-            e.stopPropagation();
-            return;
-          }
-          if (inputProps.onClick) inputProps.onClick(e)
-          setIsChecked(e?.currentTarget?.checked)
-        }}
-      />
+    return (
+      <CheckboxLabel htmlFor={id} variant={variant}>
+        <CheckboxInput
+          id={id}
+          disabled={disabled}
+          aria-label={ariaLabel}
+          ref={ref}
+          {...inputProps}
+          readOnly={readOnly}
+          onClick={(e) => {
+            if (disabled || readOnly) {
+              e.preventDefault();
+              e.stopPropagation();
+              return;
+            }
+            if (inputProps.onClick) inputProps.onClick(e);
+            setIsChecked(e?.currentTarget?.checked);
+          }}
+        />
         <CheckboxIconWrapper
-          checked={typeof inputProps.checked !== 'undefined' ? inputProps.checked : isChecked}
+          checked={
+            typeof inputProps.checked !== 'undefined'
+              ? inputProps.checked
+              : isChecked
+          }
           dir={dir}
           label={label}
         />
-    </CheckboxLabel>);
-});
+      </CheckboxLabel>
+    );
+  }
+);
 
 export default Checkbox;